1. removed redundant casts (1438828) 2. Applied patch to sort macro folders at top 3. Applied sourcecode formatting to some classes.
git-svn-id: svn+ssh://svn.code.sf.net/p/jedit/svn/jEdit/trunk@5354 6b1eeb88-9816-0410-afa2-b43733a0f04e

+++ bsh/Primitive.java
@@ -864,7 +864,7 @@ public final class Primitive implements ParserConstants, java.io.Serializable
 
 		// non boolean primitive, get appropriate flavor of zero
 		try {
-			return new Primitive((int)0).castToType( type, Types.CAST );
+			return new Primitive(0).castToType( type, Types.CAST );
 		} catch ( UtilEvalError e ) {
 			throw new InterpreterError( "bad cast" );
 		}
